[Refactor] 환경변수 개선 및 프로퍼티 관련 패키지 이동 (#35)
* refac: 환경별 공통으로 사용되는 환경변수 추출

* refac: 인증 피처 프로퍼티 클래스 위치 이동

* refac: `@ConfigurationPropertiesScan`으로 감지하고 있으므로 프로퍼티를 빈으로 등록하는 클래스 제거

* refac: spring.datasource.url 베이스 데이터 수정
