+++ .gitignore
@@ -66,9 +66,5 @@ bin/
 ### Some additional ignores
 *.DS_Store
 
-
-### Application Properties ###
-src/main/resources/application.yml
-
 # SSH Tunneling
 *.pem
