+++ src/main/java/com/evenly/took/feature/auth/application/JwtTokenProvider.java
@@ -6,9 +6,9 @@ import java.util.Date;
 
 import org.springframework.stereotype.Component;
 
+import com.evenly.took.feature.auth.config.properties.TokenProperties;
 import com.evenly.took.feature.auth.exception.AuthErrorCode;
 import com.evenly.took.feature.common.exception.TookException;
-import com.evenly.took.global.config.properties.auth.TokenProperties;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
