+++ src/main/java/com/evenly/took/feature/auth/client/apple/AppleAuthCodeRequestUrlProvider.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Component;
 
 import com.evenly.took.feature.auth.client.AuthCodeRequestUrlProvider;
 import com.evenly.took.feature.auth.client.apple.dto.request.AppleAuthRequest;
+import com.evenly.took.feature.auth.config.properties.AppleProperties;
+import com.evenly.took.feature.auth.config.properties.AppleUrlProperties;
 import com.evenly.took.feature.auth.domain.OAuthType;
-import com.evenly.took.global.config.properties.auth.AppleProperties;
-import com.evenly.took.global.config.properties.auth.AppleUrlProperties;
 
 import lombok.RequiredArgsConstructor;
