+++ src/main/java/com/evenly/took/feature/auth/client/kakao/KakaoAuthCodeRequestUrlProvider.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import com.evenly.took.feature.auth.client.AuthCodeRequestUrlProvider;
+import com.evenly.took.feature.auth.config.properties.KakaoProperties;
 import com.evenly.took.feature.auth.domain.OAuthType;
-import com.evenly.took.global.config.properties.auth.KakaoProperties;
 
 import lombok.RequiredArgsConstructor;
