+++ src/main/java/com/evenly/took/feature/auth/client/kakao/KakaoTokenProvider.java
@@ -10,7 +10,7 @@ import org.springframework.web.client.RestClient;
 import com.evenly.took.feature.auth.client.kakao.dto.request.KakaoTokenRequest;
 import com.evenly.took.feature.auth.client.kakao.dto.response.KakaoTokenResponse;
 import com.evenly.took.feature.auth.client.kakao.error.KakaoTokenProviderErrorHandler;
-import com.evenly.took.global.config.properties.auth.KakaoProperties;
+import com.evenly.took.feature.auth.config.properties.KakaoProperties;
 
 @Component
 public class KakaoTokenProvider {
