+++ src/main/java/com/evenly/took/feature/auth/client/kakao/KakaoUserInfoProvider.java
@@ -9,7 +9,7 @@ import org.springframework.web.client.RestClient;
 
 import com.evenly.took.feature.auth.client.kakao.dto.response.KakaoUserResponse;
 import com.evenly.took.feature.auth.client.kakao.error.KakaoUserInfoProviderErrorHandler;
-import com.evenly.took.global.config.properties.auth.KakaoProperties;
+import com.evenly.took.feature.auth.config.properties.KakaoProperties;
 
 @Component
 public class KakaoUserInfoProvider {
