--- src/main/java/com/evenly/took/global/config/properties/PropertiesConfig.java
@@ -1,13 +1,0 @@
-package com.evenly.took.global.config.properties;
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-import com.evenly.took.global.config.properties.auth.TokenProperties;
-
-@EnableConfigurationProperties({
-	TokenProperties.class,
-})
-@Configuration
-public class PropertiesConfig {
-}
