+++ src/test/java/com/evenly/took/feature/auth/client/apple/AppleTokenProviderTest.java
@@ -15,8 +15,8 @@ import org.springframework.http.MediaType;
 import org.springframework.web.client.RestClient;
 
 import com.evenly.took.feature.auth.client.apple.error.AppleTokenProviderErrorHandler;
-import com.evenly.took.global.config.properties.auth.AppleProperties;
-import com.evenly.took.global.config.properties.auth.AppleUrlProperties;
+import com.evenly.took.feature.auth.config.properties.AppleProperties;
+import com.evenly.took.feature.auth.config.properties.AppleUrlProperties;
 import com.evenly.took.global.service.MockTest;
 
 class AppleTokenProviderTest extends MockTest {
