+++ src/test/java/com/evenly/took/feature/auth/client/kakao/KakaoTokenProviderTest.java
@@ -12,12 +12,14 @@ import org.springframework.boot.test.autoconfigure.web.client.RestClientTest;
 
 import com.evenly.took.feature.auth.client.kakao.dto.response.KakaoTokenResponse;
 import com.evenly.took.feature.auth.client.kakao.error.KakaoTokenProviderErrorHandler;
+import com.evenly.took.feature.auth.config.properties.KakaoProperties;
 import com.evenly.took.feature.auth.exception.AuthErrorCode;
 import com.evenly.took.feature.common.exception.TookException;
-import com.evenly.took.global.config.properties.auth.KakaoProperties;
 import com.evenly.took.global.service.BaseRestClientTest;
 
-@RestClientTest({KakaoTokenProvider.class, KakaoTokenProviderErrorHandler.class})
+@RestClientTest({KakaoTokenProvider.class,
+	KakaoTokenProviderErrorHandler.class,
+	KakaoProperties.class})
 class KakaoTokenProviderTest extends BaseRestClientTest {
 
 	KakaoTokenProvider kakaoTokenProvider;
