+++ src/test/java/com/evenly/took/global/service/BaseRestClientTest.java
@@ -17,11 +17,10 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestClient;
 
-import com.evenly.took.feature.auth.config.AuthConfig;
 import com.evenly.took.global.config.client.ClientConfig;
 
 @ActiveProfiles("test")
-@Import({ClientConfig.class, AuthConfig.class})
+@Import({ClientConfig.class})
 public abstract class BaseRestClientTest {
 
 	@Autowired
